src/logging_service/requirements.txt
confluent-kafka==2.6.0
grpcio
grpcio-tools
protobuf